Adding merge function as an optional input for KeyZipper (#76)
Summary:
Pull Request resolved: https://github.com/pytorch/data/pull/76

Fixes #44 by adding merge function as an optional input for `KeyZipper`. It allows user to specify how the items yielded from the DataPipes will be combined before they are yielded.

This PR changes the default behavior when `keep_key=True` and that may be BC breaking.
Previously, when `keep_key=True`, the output will be `(key, item1, item2)`, now it will default to `(key, (item1, item2)`. It is possible to leave the default behavior unchanged.

Note that `KeyZipper` may be renamed to `IterZipper` in #50 to provide users with better clarity.

Test Plan: Imported from OSS

Reviewed By: ejguan

Differential Revision: D31820203

Pulled By: NivekT

fbshipit-source-id: 670941e03595f0b0013c1162cfe1c22ec3ce6064