add RarArchiveReader (#41)
Summary:
This works as expected, but will be harder to test than the other archive readers. The problem is that since the `rarfile` package can only read rar archives but not write them, we cannot generate the test data at runtime. One option would be to put fixed data in say `test/assets/test_rar.rar` and use this.

Pull Request resolved: https://github.com/pytorch/data/pull/41

Reviewed By: NivekT

Differential Revision: D32831804

Pulled By: ejguan

fbshipit-source-id: fc8bc488be9f829b648e89f978d2b811194ec5d4