Switching to use FileOpener instead of FileLoader (#148)
Summary:
Pull Request resolved: https://github.com/pytorch/data/pull/148

This PR is dependent on https://github.com/pytorch/pytorch/pull/70367, which renames `FileLoader` to `FileOpener` in PyTorch. Core library. Then, this PR switches all `TorchData` usage of `FileLoader` to `FileOpener`, as well as renaming `IoPathFileLoader` to `IoPathFileOpener` (the functional name changes as well). Downstream users of those DataPipes please be aware of that.

Fixes #103.

Reviewed By: ejguan

Differential Revision: D33368187

fbshipit-source-id: aabb545c002c04f1fd305fd64e9982088d6a81a8